+++ sdk/cosmos/azure_data_cosmos/src/routing/mod.rs
@@ -7,6 +7,7 @@ pub mod container_cache;
 pub mod global_endpoint_manager;
 pub mod global_partition_endpoint_manager;
 pub mod location_cache;
+pub mod metadata_hedging;
 pub mod partition_key_range;
 pub mod partition_key_range_cache;
 pub mod range;